Revert "Revert "feature: commission_rate_in_basis_points (#9822)" (#9847)"
This reverts commit 4382931e57ff7757d017dc31b0486dc258bfe511.

+++ src/flamenco/stakes/fd_vote_stakes_private.h
@@ -32,7 +32,7 @@ struct index_ele {
   uint        prev_multi;
   uint        next_multi;
   ushort      refcnt;
-  uchar       commission_t_2;
+  ushort      commission_t_2;
   uchar       exists_t_2;
 };
 typedef struct index_ele index_ele_t;
